requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
crawl4ai>=0.1.0
deep-translator>=1.11.0
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.0
scikit-learn>=1.0.0